refactor: MachineType model
